video_core: Refactor and render target swizzles
