+++ src/shader_recompiler/frontend/translate/vector_alu.cpp
@@ -479,10 +479,11 @@ void Translator::V_ADD_F32(const GcnInst& inst) {
 void Translator::V_CVT_OFF_F32_I4(const GcnInst& inst) {
     const IR::U32 src0{GetSrc(inst.src[0])};
     const IR::VectorReg dst_reg{inst.dst[0].code};
-    ir.SetVectorReg(
-        dst_reg,
-        ir.FPMul(ir.ConvertUToF(32, 32, ir.ISub(ir.BitwiseAnd(src0, ir.Imm32(0xF)), ir.Imm32(8))),
-                 ir.Imm32(1.f / 16.f)));
+    ASSERT(src0.IsImmediate());
+    static constexpr std::array IntToFloat = {
+        0.0f,     0.0625f,  0.1250f,  0.1875f,  0.2500f,  0.3125f,  0.3750f,  0.4375f,
+        -0.5000f, -0.4375f, -0.3750f, -0.3125f, -0.2500f, -0.1875f, -0.1250f, -0.0625f};
+    ir.SetVectorReg(dst_reg, ir.Imm32(IntToFloat[src0.U32()]));
 }
 
 void Translator::V_MED3_F32(const GcnInst& inst) {
